--- Scripts/README.MD
@@ -1,9 +1,0 @@
-Place this folder into your user's home directory. 
-
-The default username is fpvout. If you use a different username, you will need to modify the scripts so that they point to your home folder.
-
-## Credit
-
-The piShutdown.py script was written by Core Electronics. It is included in DigiView with their permission.
-
-Instructions for use of the script are [here](https://core-electronics.com.au/tutorials/how-to-make-a-safe-shutdown-button-for-raspberry-pi.html)
