--- Scripts/fpv-c/README.MD
@@ -1,1 +1,0 @@
-fpv-c was written by [FrankPetrilli](https://github.com/FrankPetrilli)
