--- Services/GoggleConnect.service
@@ -1,11 +1,0 @@
-[Unit] 
-Description=GoggleConnect 
-StopWhenUnneeded=true 
-
-[Service] 
-Type=Oneshot 
-User=fpvout 
-ExecStart=/home/fpvout/Scripts/fpvout-start.sh 
-
-[Install] 
-WantedBy=multi-user.target
