--- Services/README.MD
@@ -1,3 +1,0 @@
-Place the contents of this folder into the following directory:
-
-/etc/systemd/system
