--- UDEV Rules/README.MD
@@ -1,1 +1,0 @@
-Place the contents of this folder into /etc/udev/rules.d
