+++ hello_video.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+sudo mv video.c /opt/vc/src/hello_pi/hello_video/video.c
+sudo /opt/vc/src/hello_pi/rebuild.sh
