--- hello_video/README.MD
@@ -1,5 +1,0 @@
-This version of hello_video is modified to allow for live video streaming via H.264. 
-
-Thanks to [befinitiv](https://github.com/befinitiv)
-
-The source repository for this modified script, along with instructions for using it is [here](https://github.com/befinitiv/hello_video)
